bugfix: вынести интерфейс из класса Main
